Ease the work around Markdown
Added a VSCode task that installs two extensions which get rid of the tedious task of manually updating the Table of Contents and Section Numbers.

A README file has been created that contains an in-depth guide for the Markdown All in One extension.

Moreover, all the documents' Table of Contents have been updated to day.